Add cookiecutter-pypackage-rust-cross-platform-publish
This is a project I've been working on for a few weeks now. There is still some work to be done, but I plan to support it, and think most of the major technical issues have been resolved. An example of the output can be seen at https://github.com/mckaymatt/trust_pypi_example/releases